+++ src/mspparser.hpp
@@ -39,6 +39,12 @@ namespace hf {
             
             // Number of EEPROM reserved slots for parameters
             static const int PARAMETER_SLOTS = 150;
+            uint8_t MISSION_COMMANDS[13] = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13};
+            uint8_t _lastCommandData = 0;
+            
+        protected:
+
+          bool incomingMission = false;
 
         private:
 
@@ -46,7 +52,6 @@ namespace hf {
             static const int OUTBUF_SIZE = 128;
 
             int EEPROMindex = PARAMETER_SLOTS;
-            bool incomingMission = false;
 
             typedef enum serialState_t {
                 IDLE,
@@ -183,15 +188,23 @@ namespace hf {
 
             void processMissionCommand(uint8_t command)
             {
-                if (incomingMission && command != 23)
+                if (isMissionCommand(command))
                 {
-                    EEPROM.write(EEPROMindex, command);
-                    EEPROMindex += 1;
+                    int commandData = -1;
                     if (command != 1 && command != 2 && command != 3)
                     {
-                        uint8_t commandData = readCommandData();
-                        EEPROM.write(EEPROMindex, commandData);
+                        commandData = readCommandData();
+                        _lastCommandData = (uint8_t)commandData;
+                    }
+                    if (incomingMission)
+                    {
+                        EEPROM.write(EEPROMindex, command);
                         EEPROMindex += 1;
+                        if (commandData!=-1)
+                        {
+                          EEPROM.write(EEPROMindex, (uint8_t)commandData);
+                          EEPROMindex += 1;
+                        }
                     }
                 }
             }
@@ -199,6 +212,15 @@ namespace hf {
             uint8_t readCommandData(void)
             {
                 return _inBuf[_inBufIndex++] & 0xff;
+            }
+
+            bool isMissionCommand(uint8_t command)
+            {
+                for(int i=0; i<sizeof(MISSION_COMMANDS)/sizeof(MISSION_COMMANDS[0]); ++i)
+                {
+                    if (command == MISSION_COMMANDS[i]) return true;
+                }
+                return false;
             }
 
         protected: