+++ src/pidcontrollers/level.hpp
@@ -53,11 +53,22 @@ namespace hf {
             
             float _demandsToAngle;
             float _demandsToRate;
-
+            
+            void computeReferenceDemands(float _demands[2], state_t &  state, demands_t & demands)
+            {
+                if (state.executingMission || state.executingStack)
+                {
+                    _demands[0] = demands.setpointAngle[0];
+                    _demands[1] = demands.setpointAngle[1];
+                } else {
+                    _demands[0] = demands.roll * _demandsToAngle;
+                    _demands[1] = demands.pitch * _demandsToAngle;
+                }
+            }
 
         public:
 
-            Level(float rollLevelP, float pitchLevelP, float maxAngle = 30)
+            Level(float rollLevelP, float pitchLevelP, float maxAngle = 45, float demandsToRate = 6.0)
             {
                 PTerms[0] = rollLevelP;
                 PTerms[1] = pitchLevelP;
@@ -67,7 +78,7 @@ namespace hf {
                 // Since we work in radians:
                 // _demandsToAngle = (maxAngle*PI/180) * 2
                 _demandsToAngle = maxAngle * 2 * M_PI / 180.0f;
-                // _demandsToRate = demandsToRate;
+                _demandsToRate = demandsToRate;
                 // _lpfRoll.init();
                 // _lpfPitch.init();
             }
@@ -81,14 +92,15 @@ namespace hf {
                 _demandsToRate = demandsToRate;
             }
 
-            bool modifyDemands(eskf_state_t & state, demands_t & demands, float currentTime)
+            bool modifyDemands(state_t & state, demands_t & demands, float currentTime)
             {
                 (void)currentTime;
 
-                float _demands[2] = {demands.roll, demands.pitch};
+                float _demands[2];
+                computeReferenceDemands(_demands, state, demands);
                 for (int axis=0; axis<2; ++axis)
                 {
-                  float error = _demands[axis] * _demandsToAngle - state.eulerAngles[axis];
+                  float error = _demands[axis] - state.UAVState->eulerAngles[axis];
                   // 0.0175 rad ~ 1 degree
                   //error = fabs(error) > 0.0175 ? error : 0.0;
                   // if(axis==0) {
@@ -105,12 +117,12 @@ namespace hf {
                   // }
 
                   // XXX Debug 
-                  if (fabs(error) > 0.0175) {
-                    pinMode(25, OUTPUT);
-                    digitalWrite(25, LOW);
-                  } else {
-                    digitalWrite(25, HIGH);
-                  }
+                  // if (fabs(error) > 0.0175) {
+                  //   pinMode(25, OUTPUT);
+                  //   digitalWrite(25, LOW);
+                  // } else {
+                  //   digitalWrite(25, HIGH);
+                  // }
                   
                   float FF = 0;
                   if (fabs(_demands[axis]) > FF_THRESHOLD) 
@@ -119,8 +131,8 @@ namespace hf {
 
                 }
                 // Output of Level controller should be the desired rates
-                demands.roll = _demands[0]/_demandsToRate;
-                demands.pitch = _demands[1]/_demandsToRate;
+                demands.roll = _demands[0] / _demandsToRate;
+                demands.pitch = _demands[1] / _demandsToRate;
                 // demands.roll = _lpfRoll.update(_demands[0]/_demandsToRate);
                 // demands.pitch = _lpfPitch.update(_demands[1]/_demandsToRate);